Slurm for ensemble interaction
